Move get_tables_names_user() and get_tablename_tuple() from easyaccess.py to dbdo_utils.py
